+++ modules/update-center/locals-tags.tf
@@ -5,4 +5,9 @@ locals {
   } : {}
 
   tags = merge(local.default_tags, var.extra_tags)
+
+  # To avoid perpetual drift
+  mc_tags = {
+    for key, value in local.tags : lower(key) => value
+  }
 }